pyproject.toml
[tool.poetry]
name = "mailbot"
version = "0.1.0"
description = ""
authors = ["Terry <78312274+TerryEmeigh@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.182.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"
dotenv = "^0.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
